+++ src/components/SkeletonPage/index.ts
@@ -1,0 +1,4 @@
+import SkeletongBodyText from './SkeletonPage';
+
+export {Props} from './SkeletonPage';
+export default SkeletongBodyText;